refactor: component 구조로 리팩토링
