+++ src/js/components/lines/linesTemplate.js
