+++ src/js/components/login/Login.js
@@ -1,0 +1,1 @@
+class Login {}
