+++ src/js/components/login/loginTemplate.js
@@ -48,7 +48,7 @@ const loginTemplate = () => {
             </div>
             <p class="text-gray-700 pl-2">
             아직 회원이 아니신가요?
-            <a href="/pages/signup.html">회원가입</a>
+            <a href="/signup">회원가입</a>
             </p>
         </form>
     </div>
