+++ src/js/components/sections/sectionsTemplate.js
