+++ src/js/components/signup/Signup.js
@@ -1,0 +1,1 @@
+class SignUp {}
