+++ src/js/components/stations/Stations.js
@@ -1,0 +1,1 @@
+class Stations {}
